+++ trionyx/quickstart/reusable_app/[app_name]/apps.py
@@ -1,0 +1,9 @@
+"""[[verbose_name]] app configuration"""
+from trionyx.trionyx.apps import BaseConfig
+
+
+class Config(BaseConfig):
+    """[[verbose_name]] configuration"""
+
+    name = 'apps.[[name]]'
+    verbose_name = '[[verbose_name]]'