+++ trionyx/quickstart/reusable_app/[app_name]/models.py
@@ -1,0 +1,4 @@
+"""App models"""
+from trionyx import models  # noqa F401
+
+# Create you models here (models.BaseModel)