+++ trionyx/quickstart/reusable_app/app/__init__.py
@@ -1,0 +1,4 @@
+"""Reusable app, test app"""
+from app.celery import app as celery_app
+
+__all__ = ['celery_app']