+++ trionyx/quickstart/reusable_app/app/celery.py
@@ -1,0 +1,15 @@
+"""Reusable app, Celery app"""
+import os
+from celery import Celery
+
+# set the default Django settings module for the 'celery' program.
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'app.settings')
+
+# Create celery app
+app = Celery('trionyx')
+
+# Get celery config from django settings
+app.config_from_object('django.conf:settings', namespace='CELERY')
+
+# Load task modules from all registered Django app configs.
+app.autodiscover_tasks()